backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
boto3==1.28.85
botocore==1.31.85
cachetools==5.3.2
-e git+ssh://git@github.com/ibrahimcrowdbotics/lawrence-west-app-44469.git@745989a52f46dea86fb84d0b8496a540e20f52d4#egg=cb_django_contact_us&subdirectory=backend/modules/django_contact_us
-e git+ssh://git@github.com/ibrahimcrowdbotics/lawrence-west-app-44469.git@745989a52f46dea86fb84d0b8496a540e20f52d4#egg=cb_terms_and_conditions&subdirectory=backend/modules/terms_and_conditions
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
dj-rest-auth==5.0.1
Django==3.2.23
django-allauth==0.54.0
django-autoslug==1.9.9
django-bootstrap4==22.3
django-cities-light==3.10.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-import-export==3.3.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.5
et-xmlfile==1.1.0
exceptiongroup==1.2.0
facebook-sdk==3.1.0
factory-boy==3.3.0
Faker==22.0.0
filelock==3.13.1
google-api-core==2.15.0
google-auth==2.22.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
importlib-resources==6.1.1
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
Pillow==10.0.1
pipenv==2023.11.15
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
pluggy==1.3.0
progressbar2==4.3.2
proto-plus==1.23.0
protobuf==4.25.1
psycopg2==2.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.2
python-dateutil==2.8.2
python-http-client==3.3.7
python-utils==3.8.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.16.2
rsa==4.9
s3transfer==0.7.0
sendgrid==6.11.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tablib==3.5.0
tomli==2.0.1
typing_extensions==4.9.0
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==1.26.18
virtualenv==20.25.0
waitress==2.1.2
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.17.0
